api/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-filter==23.2
idna==3.4
oauthlib==3.2.2
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pydantic==1.10
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
attrs==23.1.0
drf-spectacular==0.26.3
jsonschema==4.17.3
pyrsistent==0.19.3
PyYAML==6.0
uritemplate==4.1.1
amqp==5.1.1
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.3.1
marshmallow==3.19.0
prompt-toolkit==3.0.39
pydantic_core==2.3.0
redis==4.6.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6